Update the "Retire an application" documentation
Reordered items such that:

- Unpublishing routes is done first (since that typically involves
  a lot of planning and pre-requisite work)
- Destructive actions are done last
- All the "raise a PR" actions are grouped together

Also clarified that the Signon steps need running in Integration
as well as Production, and made it clearer that the Release app
step is a destructive action (it isn't "Mark as retired" - it's
deleting data from the Release database). Finally, removed the
reference to dropping a database, since this is handled automatically
when deleting the corresponding Terraform. Instead I've added a
general note to remove any manually-managed resources from AWS.