Update 'Retire a repo' doc around updating dev docs
We have a check that ensures all repos are listed in dev docs so it’s fair to
assume this will always be the case.

We no longer mark repos as retired with `retired: true` as they were included
in the Dev docs search results making the search results less relevant and
harder to find information.