Highlight that GitHub config managed via Terraform needs care
This was documented but easily overlooked if you're just skimming
the page (one might assume that "Remove GitHub configuration for
the repo" is some low-risk clickops in the GitHub UI, for example).
It's quite surprising behaviour that we manage our GitHub repos
externally (via Terraform) and that if we don't log into the
Terraform console and run some steps manually, our repo that we're
trying to carefully archive will actually be deleted!

Until this workaround is tackled (in https://github.com/alphagov/govuk-infrastructure/issues/1865),
let's make it clearer what the risk/impact is of failing to follow
the step, and also signpost it from the canonical "retire an app"
doc.

Related Trello: https://trello.com/c/MiBvG4Xs/