Update 'Configure a new GOV.UK repository'
The steps were not correct as GitHub terraform relies on repos to be tagged to
`govuk` topic:
https://github.com/alphagov/govuk-infrastructure/blob/478daa1e6069701dbb4cdd5f559a647fabb58ba1/terraform/deployments/github/main.tf#L46
The topic is now applied in terraform.

This replaces the verbose, repo specific guidance with a link to the the
updated guidance in the repo's README, which will make it easier to keep up to
date.