Make "Sidekiq Admin" permission open to interpretation
Whitehall has a "Sidekiq Admin" permission.
Publisher has a "sidekiq_admin" permission (so chosen to follow the same naming convention of the other permissions in Publisher).
I don't think that's an issue - but we do need to make it clearer that the name of the permission may vary. This commit does that.