Create a new app: include boilerplate files in the written document
The document previously linked out to examples of the relevant files in live
repositories. This could cause future apps to include features or logic that
are only relevant to the linked example. To avoid that, we provide boilerplate
file content that can included in the repository as a starting point.