Remove documentation on suggested related links
Suggested related links have been broken for a few years and the quality
has been atrophying. We have removed the links from the public facing
site and are in the process of removing the links from the content-store

We are have removed the developer docs on the suggested links since they
are being removed to keep the documentation clean and up to date.